refactor: reorganize Quiz component imports and remove unused components, enhancing code clarity and maintainability
